backend/requirements.txt
fastapi
pydantic
sqlalchemy
sqlmodel
PyJWT
uvicorn
dotenv
pydantic[email]
pwdlib
pwdlib[argon2]